Support custom extension matching
This commit adds support for customising extension matching.  This can
be used to install releases that are in alternative file formats, or
lack a file extension entirely e.g. pure binaries.

When a download release asset is not an archive it is copied into the
tool cache rather than being extracted to there.  The copied asset may
optionally be renamed and chmod'd if desired.

Also adds support for extracting release archives that are .tar.bz2
format.